Fixing imports

+++ examples/evm/stablecoin_collateral/src/Interfaces/IBorrowVenueAdapter.sol
@@ -1,9 +1,12 @@
+// SPDX-License-Identifier: MIT
 pragma solidity ^0.8.24;
 
 interface IBorrowVenueAdapter {
     // Collateral
     function supplyCollateral(address collateral, uint256 amount, bytes calldata data) external;
-    function withdrawCollateral(address collateral, uint256 amount, bytes calldata data) external returns (uint256 withdrawn);
+    function withdrawCollateral(address collateral, uint256 amount, bytes calldata data)
+        external
+        returns (uint256 withdrawn);
     function collateralBalance(address collateral) external view returns (uint256);
 
     // Debt
@@ -11,5 +14,8 @@ interface IBorrowVenueAdapter {
     function repay(address debtAsset, uint256 amount, bytes calldata data) external returns (uint256 repaid);
     function debtBalance(address debtAsset) external view returns (uint256);
 
-    function getRiskSignals(address debtAsset) external view  returns (bool hasApr, uint aprBps, bool haveHf, uint hfBps);
+    function getRiskSignals(address debtAsset)
+        external
+        view
+        returns (bool hasApr, uint256 aprBps, bool haveHf, uint256 hfBps);
 }
